requirements.txt
sqlalchemy
snowflake-sqlalchemy
